+++ settings.php
@@ -1454,8 +1454,7 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
             'theme_boost_union',
             ['settingname' => $locallogincorenotesettingname, 'url' => $locallogincorenoteurl->out()],
             true
-        ), \core\output\notification::NOTIFY_INFO);
-        $locallogincorenote->set_show_closebutton(false);
+        ), \core\output\notification::NOTIFY_INFO, false);
         $description .= $OUTPUT->render($locallogincorenote);
         $localloginurl = new core\url('/theme/boost_union/locallogin.php');
         $localloginnotification = new \core\output\notification(get_string(
@@ -1463,8 +1462,7 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
             'theme_boost_union',
             ['url' => $localloginurl],
             true
-        ), \core\output\notification::NOTIFY_WARNING);
-        $localloginnotification->set_show_closebutton(false);
+        ), \core\output\notification::NOTIFY_WARNING, false);
         $description .= $OUTPUT->render($localloginnotification);
         $setting = new admin_setting_configselect($name, $title, $description, THEME_BOOST_UNION_SETTING_SELECT_YES, $yesnooption);
         $tab->add($setting);
@@ -1624,8 +1622,7 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
             'theme_boost_union',
             ['url' => $idplogincorenoteurl->out()],
             true
-        ), \core\output\notification::NOTIFY_INFO);
-        $idplogincorenote->set_show_closebutton(false);
+        ), \core\output\notification::NOTIFY_INFO, false);
         $description .= $OUTPUT->render($idplogincorenote);
         $setting = new admin_setting_configselect($name, $title, $description, THEME_BOOST_UNION_SETTING_SELECT_YES, $yesnooption);
         $tab->add($setting);
@@ -1885,8 +1882,7 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
             'theme_boost_union',
             ['settingname' => $selfregistrationlogincorenotesettingname, 'url' => $selfregistrationlogincorenoteurl->out()],
             true
-        ), \core\output\notification::NOTIFY_INFO);
-        $selfregistrationlogincorenote->set_show_closebutton(false);
+        ), \core\output\notification::NOTIFY_INFO, false);
         $description .= $OUTPUT->render($selfregistrationlogincorenote);
         $setting = new admin_setting_configselect($name, $title, $description, THEME_BOOST_UNION_SETTING_SELECT_YES, $yesnooption);
         $tab->add($setting);
@@ -2034,8 +2030,7 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
             'theme_boost_union',
             ['settingname' => $guestlogincorenotesettingname, 'url' => $guestlogincorenoteurl->out()],
             true
-        ), \core\output\notification::NOTIFY_INFO);
-        $guestlogincorenote->set_show_closebutton(false);
+        ), \core\output\notification::NOTIFY_INFO, false);
         $description .= $OUTPUT->render($guestlogincorenote);
         $setting = new admin_setting_configselect($name, $title, $description, THEME_BOOST_UNION_SETTING_SELECT_YES, $yesnooption);
         $tab->add($setting);
@@ -3177,9 +3172,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
                     ['url' => $mobilecssurl->out(), 'value' => $CFG->mobilecssurl]
                 ) . ' ' .
                     get_string('mobilecss_donotchange', 'theme_boost_union'),
-                \core\output\notification::NOTIFY_WARNING
+                \core\output\notification::NOTIFY_WARNING,
+                false
             );
-            $mobilescssnotification->set_show_closebutton(false);
             $description .= $OUTPUT->render($mobilescssnotification);
 
             // Otherwise, we just add a note to the description.
@@ -3191,9 +3186,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
                     ['url' => $mobilecssurl->out()]
                 ) . ' ' .
                     get_string('mobilecss_donotchange', 'theme_boost_union'),
-                \core\output\notification::NOTIFY_INFO
+                \core\output\notification::NOTIFY_INFO,
+                false
             );
-            $mobilescssnotification->set_show_closebutton(false);
             $description .= $OUTPUT->render($mobilescssnotification);
         }
         // Using admin_setting_scsscode is not 100% right here as this setting does not support SCSS.
@@ -3494,9 +3489,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/blockregionsintro';
         $blockregionsintro = new \core\output\notification(
             get_string('blockregionsheading_desc', 'theme_boost_union'),
-            \core\output\notification::NOTIFY_INFO
+            \core\output\notification::NOTIFY_INFO,
+            false
         );
-        $blockregionsintro->set_show_closebutton(false);
         $blockregionsintro->set_extra_classes(['alert-dark']);
         $description = $OUTPUT->render($blockregionsintro);
         $setting = new admin_setting_heading($name, '', $description);
@@ -3506,9 +3501,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/blockregionsheadingexperimental';
         $notification = new \core\output\notification(
             get_string('blockregionsheading_experimental', 'theme_boost_union'),
-            \core\output\notification::NOTIFY_WARNING
+            \core\output\notification::NOTIFY_WARNING,
+            false
         );
-        $notification->set_show_closebutton(false);
         $description = $OUTPUT->render($notification);
         $setting = new admin_setting_heading($name, '', $description);
         $tab->add($setting);
@@ -3589,9 +3584,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
                 $notificationurl = 'https://docs.moodle.org/en/Block_settings#Making_a_block_sticky_throughout_the_whole_site';
                 $notification = new \core\output\notification(
                     get_string('blockregionsstickyonly', 'theme_boost_union', $notificationurl),
-                    \core\output\notification::NOTIFY_INFO
+                    \core\output\notification::NOTIFY_INFO,
+                    false
                 );
-                $notification->set_show_closebutton(false);
                 $description .= '<br />' . $OUTPUT->render($notification);
             }
             $setting = new admin_setting_configmulticheckbox($name, $title, $description, [], $regions);
@@ -5712,9 +5707,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $notificationurl = 'https://github.com/moodle-an-hochschulen/moodle-theme_boost_union_snippets';
         $notification = new \core\output\notification(
             get_string('snippetsuploadedsnippetsheading_desc', 'theme_boost_union', $notificationurl),
-            \core\output\notification::NOTIFY_INFO
+            \core\output\notification::NOTIFY_INFO,
+            false
         );
-        $notification->set_show_closebutton(false);
         $description = $OUTPUT->render($notification);
         $setting = new admin_setting_heading($name, $title, $description);
         $tab->add($setting);